refactor(operator): dispatch authorization projection through provider projectors
Introduce a single PolicyProjector interface for authorization projection and dispatch projected desired state through the configured provider implementation.

Move the KAOS ConfigMap sink and broker admin sink into internal/authz/adapters so each provider owns only its required fields and helper methods.

Move broker serialization helpers from internal/aib into the adapter package, keeping internal/aib focused on the admin client.

Remove the dual provider gate booleans and the unused broker field from the generic reconciler, and return projector errors uniformly for reconcile backoff.
