+++ operator/main.go
@@ -21,6 +21,7 @@ import (
 	kaosv1alpha1 "github.com/axsaucedo/kaos/operator/api/v1alpha1"
 	"github.com/axsaucedo/kaos/operator/controllers"
 	"github.com/axsaucedo/kaos/operator/internal/aib"
+	"github.com/axsaucedo/kaos/operator/internal/authz/adapters"
 	"github.com/axsaucedo/kaos/operator/pkg/security"
 )
 
@@ -105,38 +106,37 @@ func main() {
 		os.Exit(1)
 	}
 
-	// The authorization projection controller applies the configured provider:
-	// the KAOS provider writes projected grant data into the policy ConfigMap
-	// (no broker required), while the AIB provider registers agents, mints
-	// per-agent credential Secrets, and projects services/permission sets into
-	// the broker admin API. It runs independently of the workload controllers, so
-	// a broker outage never stalls workload reconciliation. Provider "none"
-	// (default) leaves the operator without any projection.
+	// The authorization projection controller applies the configured provider.
+	// Provider "none" (default) leaves the operator without any projection.
 	if provider := security.GetConfig().AuthzProviderOrDefault(); provider != security.AuthzProviderNone {
-		brokerProjection := provider == security.AuthzProviderAIB
-		policyDataProjection := provider == security.AuthzProviderKAOS
-
-		var admin controllers.AIBAdmin
-		if brokerProjection {
-			admin = aib.New(
-				os.Getenv("AIB_ADMIN_URL"),
-				getEnvWithDefault("AIB_PRINCIPAL", "kaos-operator"),
-				getEnvWithDefault("AIB_PRINCIPAL_HEADER", "X-Remote-User"),
-				getDurationWithDefault("AIB_REQUEST_TIMEOUT", 10*time.Second),
-			)
+		var projector controllers.PolicyProjector
+		switch provider {
+		case security.AuthzProviderKAOS:
+			projector = &adapters.ConfigMapProjector{
+				Client:    mgr.GetClient(),
+				Name:      os.Getenv("AUTHZ_POLICY_CONFIGMAP_NAME"),
+				Namespace: os.Getenv("AUTHZ_POLICY_CONFIGMAP_NAMESPACE"),
+				JWKSURI:   security.GetConfig().AuthzJWKSURI(),
+			}
+		case security.AuthzProviderAIB:
+			projector = &adapters.BrokerProjector{
+				Client: mgr.GetClient(),
+				Scheme: mgr.GetScheme(),
+				AIB: aib.New(
+					os.Getenv("AIB_ADMIN_URL"),
+					getEnvWithDefault("AIB_PRINCIPAL", "kaos-operator"),
+					getEnvWithDefault("AIB_PRINCIPAL_HEADER", "X-Remote-User"),
+					getDurationWithDefault("AIB_REQUEST_TIMEOUT", 10*time.Second),
+				),
+				SecretPrefix: getEnvWithDefault("SECURITY_AGENT_AUTH_CREDENTIAL_SECRET_PREFIX", "kaos-aib"),
+				Prune:        getBoolWithDefault("AUTHZ_PROJECTION_PRUNE_ENABLED", true),
+			}
 		}
 		if err = (&controllers.AuthzProjectionReconciler{
-			Client:                   mgr.GetClient(),
-			Scheme:                   mgr.GetScheme(),
-			AIB:                      admin,
-			Namespaces:               splitCSV(os.Getenv("AUTHZ_PROJECTION_NAMESPACES")),
-			SecretPrefix:             getEnvWithDefault("SECURITY_AGENT_AUTH_CREDENTIAL_SECRET_PREFIX", "kaos-aib"),
-			Prune:                    getBoolWithDefault("AUTHZ_PROJECTION_PRUNE_ENABLED", true),
-			BrokerProjection:         brokerProjection,
-			PolicyDataProjection:     policyDataProjection,
-			PolicyConfigMapName:      os.Getenv("AUTHZ_POLICY_CONFIGMAP_NAME"),
-			PolicyConfigMapNamespace: os.Getenv("AUTHZ_POLICY_CONFIGMAP_NAMESPACE"),
-			AgentJWKSURI:             security.GetConfig().AuthzJWKSURI(),
+			Client:     mgr.GetClient(),
+			Scheme:     mgr.GetScheme(),
+			Namespaces: splitCSV(os.Getenv("AUTHZ_PROJECTION_NAMESPACES")),
+			Projector:  projector,
 		}).SetupWithManager(mgr); err != nil {
 			setupLog.Error(err, "unable to create controller", "controller", "AuthzProjection")
 			os.Exit(1)
